+++ tinper-bee/bee-datepicker/src/rc-calendar/Calendar.jsx
@@ -94,6 +94,8 @@ class Calendar extends React.Component {
   onKeyDown = (event) => {
     if (event.target.nodeName.toLowerCase() === 'input') {
       return undefined;
+    }else{
+      this.props.onKeyDown&&this.props.onKeyDown(event);
     }
     const keyCode = event.keyCode;
     // mac
@@ -152,9 +154,6 @@ class Calendar extends React.Component {
           });
         }
         event.preventDefault();
-        return 1;
-      default:
-        this.props.onKeyDown(event);
         return 1;
     }
   }