+++ tinper-bee/bee-table/demo/demolist/Demo1001.js
@@ -2,7 +2,8 @@
 *
 * @title 拖拽改变列顺序
 * @parent 列操作-拖拽 Drag
-* @description 点击列的表头，进行左右拖拽，注意：固定列不可以交换
+* @description 点击选择表头并左右拖拽，可以改变表格列顺序。onDrop方法是拖拽交换列后触发的回调函数。注意：固定列不可以交换。
+* demo1001
 */
 import React, { Component } from 'react';
 import {Icon} from "tinper-bee";
@@ -10,21 +11,21 @@ import {Icon} from "tinper-bee";
 import Table from '../../src'; 
 import dragColumn from '../../src/lib/dragColumn';
 
-const columns22 = [
+const columns = [
   {
-    title: "名字",
+    title: "订单编号",
     dataIndex: "a",
     key: "a",
     width: 100
   },
   {
-    title: "性别",
+    title: "单据日期",
     dataIndex: "b",
     key: "b",
     width: 200
   },
   {
-    title: "年龄",
+    title: "供应商",
     dataIndex: "c",
     key: "c",
     width: 200,
@@ -32,17 +33,17 @@ const columns22 = [
     sorter: (a, b) => a.c - b.c
   },
   {
-    title: "武功级别",
+    title: "联系人",
     dataIndex: "d",
     key: "d",
     width: 200,
   }
 ];
 
-const data22 = [
-  { a: "杨过", b: "男", c: 30,d:'内行', key: "2" },
-  { a: "令狐冲", b: "男", c: 41,d:'大侠', key: "1" },
-  { a: "郭靖", b: "男", c: 25,d:'大侠', key: "3" }
+const data = [
+  { a: "NU0391001", b: "2019-03-01", c: 'xx供应商',d:'Tom', key: "2" },
+  { a: "NU0391002", b: "2018-11-02", c: 'yy供应商',d:'Jack', key: "1" },
+  { a: "NU0391003", b: "2019-05-03", c: 'zz供应商',d:'Jane', key: "3" }
 ];
 
 const DragColumnTable = dragColumn(Table);
@@ -57,13 +58,20 @@ class Demo22 extends Component {
   }
  
   render() {
-    return <DragColumnTable columns={columns22} data={data22} bordered
-    
-    draggable={true} 
+    return <DragColumnTable 
+            columns={columns} 
+            data={data} 
+            bordered
+            draggable={true} 
+            onDrop ={(event,data,columns)=>{
+              console.log("--拖拽交换列后触发事件");
+              console.log("event.target:",event.target);
+              console.log("data:",data);
+              console.log("拖拽完成后的columns:",columns);
+            }}
     />;
   }
 }
+
 Demo22.defaultProps = defaultProps22;
-
-
 export default Demo22;