+++ tinper-bee/bee-transfer/src/item.js
@@ -14,18 +14,19 @@ class Item extends React.Component{
   shouldComponentUpdate(...args) {
     return PureRenderMixin.shouldComponentUpdate.apply(this, args);
   }
-  matchFilter = (text) => {
-    const { filter, filterOption, item } = this.props;
-    if (filterOption) {
-      return filterOption(filter, item);
-    }
-    return text.indexOf(filter) >= 0;
-  }
+  // matchFilter = (text) => {
+  //   const { filter, filterOption, item } = this.props;
+  //   if (filterOption) {
+  //     return filterOption(filter, item);
+  //   }
+  //   return text.indexOf(filter) >= 0;
+  // }
   render() {
-    const { render, filter, item, lazy, checked, prefixCls, onClick,renderedText,renderedEl } = this.props;
+    const { render, filter, item, lazy, checked, prefixCls, onClick,renderedText,renderedEl, showCheckbox } = this.props;
     const className = classNames({
       [`${prefixCls}-content-item`]: true,
       [`${prefixCls}-content-item-disabled`]: item.disabled,
+      [`${prefixCls}-content-item-selected`]: checked
     });
 
     const lazyProps = assign({
@@ -60,7 +61,11 @@ class Item extends React.Component{
               title={renderedText}
               onClick={item.disabled ? undefined : () => onClick(item)}
             >
+            {
+              showCheckbox?
               <Checkbox checked={checked} disabled={item.disabled} onClick={item.disabled ? undefined : () => onClick(item)}/>
+              :''
+            }
               <span>{renderedEl}</span>
             </li>
           </Lazyload>